ad-consumer/requirements.txt
kafka-python-ng
six
pymongo
gunicorn